+++ test/multidomain/embedded/1d3d/1p_richards/spatialparams_root.hh
@@ -49,7 +49,7 @@ class RootSpatialParams
     using GlobalPosition = typename Element::Geometry::GlobalCoordinate;
 
     //! Indices to access the parameters in the dgf file
-    enum DGFParamIndices {
+    enum DGFParamIndicesSurfaceModel {
         orderIdx = 0,
         rootIdIdx = 1,
         surfaceIdx = 2,
@@ -66,8 +66,8 @@ public:
     : ParentType(gridGeometry), gridData_(gridData)
     {
         porosity_ = getParam<Scalar>("Root.SpatialParams.Porosity", 0.4);
-        constantKx_ = getParam<Scalar>("SpatialParams.Kx", 5.0968e-17);
-        constantKr_ = getParam<Scalar>("SpatialParams.Kr", 2.04e-13);
+        constantKx_ = getParam<Scalar>("Root.SpatialParams.Kx", 5.0968e-17);
+        constantKr_ = getParam<Scalar>("Root.SpatialParams.Kr", 2.04e-13);
 
         const auto& gv = gridGeometry->gridView();
         radii_.resize(gv.size(0));
@@ -75,11 +75,18 @@ public:
         {
             const auto eIdx = gv.indexSet().index(element);
             auto level0element = element;
-            for(auto levelIdx = element.level(); levelIdx != 0; levelIdx--)
+            for (auto levelIdx = element.level(); levelIdx != 0; levelIdx--)
                 level0element = level0element.father();
-            const Scalar rootLength = element.geometry().volume();
-            const Scalar rootSurface = gridData_->parameters(level0element)[DGFParamIndices::surfaceIdx]/(1 << element.level());
-            radii_[eIdx] = rootSurface / rootLength / 2.0 / M_PI;
+
+            const auto& params = gridData_->parameters(level0element);
+            if (params.size() == 1) // assume only radius is given
+                radii_[eIdx] = params[0];
+            else // assume DGFParamIndicesSurfaceModel
+            {
+                const Scalar rootLength = element.geometry().volume();
+                const Scalar rootSurface = gridData_->parameters(level0element)[DGFParamIndicesSurfaceModel::surfaceIdx]/(1 << element.level());
+                radii_[eIdx] = rootSurface / rootLength / 2.0 / M_PI;
+            }
         }
     }